refactor: pull members up to parent classes (2)
